rename stream and channels
