+++ java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferInputStream.java
@@ -19,33 +19,48 @@
 
 package org.apache.fury.io;
 
-import java.nio.ByteBuffer;
-import java.nio.channels.ReadableByteChannel;
+import java.io.IOException;
+import java.io.InputStream;
 import org.apache.fury.memory.MemoryBuffer;
+import org.apache.fury.util.Preconditions;
 
-/** {@link ReadableByteChannel} based on fury {@link MemoryBuffer}. */
-public class FuryReadableByteChannel implements ReadableByteChannel {
-  private boolean open = true;
+/** InputStream based on {@link MemoryBuffer}. */
+public class MemoryBufferInputStream extends InputStream {
   private final MemoryBuffer buffer;
 
-  public FuryReadableByteChannel(MemoryBuffer buffer) {
+  public MemoryBufferInputStream(MemoryBuffer buffer) {
     this.buffer = buffer;
   }
 
-  @Override
-  public int read(ByteBuffer dst) {
-    int position = dst.position();
-    buffer.read(dst);
-    return dst.position() - position;
+  public int read() {
+    if (buffer.remaining() == 0) {
+      return -1;
+    } else {
+      return buffer.readByte() & 0xFF;
+    }
+  }
+
+  public int read(byte[] bytes, int offset, int length) throws IOException {
+    if (length == 0) {
+      return 0;
+    }
+    int size = Math.min(buffer.remaining(), length);
+    if (size == 0) {
+      return -1;
+    }
+    buffer.readBytes(bytes, offset, size);
+    return size;
   }
 
   @Override
-  public boolean isOpen() {
-    return open;
+  public long skip(long n) throws IOException {
+    Preconditions.checkArgument(n < Integer.MAX_VALUE);
+    int nbytes = (int) Math.min(n, buffer.remaining());
+    buffer.increaseReaderIndex(nbytes);
+    return nbytes;
   }
 
-  @Override
-  public void close() {
-    open = false;
+  public int available() throws IOException {
+    return buffer.remaining();
   }
 }
