+++ java/fury-core/src/test/java/org/apache/fury/io/MockWritableChannelTest.java
@@ -19,37 +19,22 @@
 
 package org.apache.fury.io;
 
+import static org.testng.Assert.assertEquals;
+
 import java.nio.ByteBuffer;
-import java.nio.channels.WritableByteChannel;
+import org.testng.annotations.Test;
 
-/**
- * A helper class to track the size of allocations. Writes to this stream do not copy or retain any
- * data, they just bump a size counter that can be later used to know exactly which data size needs
- * to be allocated for actual writing.
- */
-public class MockWritableByteChannel implements WritableByteChannel {
-  private boolean open = true;
-  private int totalBytes;
+public class MockWritableChannelTest {
 
-  @Override
-  public int write(ByteBuffer src) {
-    int remaining = src.remaining();
-    src.position(src.limit());
-    totalBytes += remaining;
-    return remaining;
-  }
-
-  public int totalBytes() {
-    return totalBytes;
-  }
-
-  @Override
-  public boolean isOpen() {
-    return open;
-  }
-
-  @Override
-  public void close() {
-    open = false;
+  @Test
+  public void testTotalBytes() {
+    try (MockWritableChannel channel = new MockWritableChannel()) {
+      channel.write(ByteBuffer.allocate(100));
+      channel.write(ByteBuffer.allocateDirect(100));
+      ByteBuffer buffer = ByteBuffer.allocate(100);
+      buffer.position(50);
+      channel.write(buffer);
+      assertEquals(channel.totalBytes(), 250);
+    }
   }
 }
